requirements.txt
ultralytics>=8.0.100
opencv-python
numpy
pillow
pandas